fix: comment node text failing to reinsert after removal
Since the type of parent node of the comment node was calculated after the comment was removed from the document, it was failing to determine the correct parent type because there was no content at the start of the comment (because the comment was no longer there).
Moving detection of the parent node type and creating the new node with the text that was in the comment before the comment is deleted ensures that content will not be lost.