feat: use redux instead of react for cms item state
This replaces the usage of `flatData` adn `setFlatData` within the `ItemDetailsPage` component and many of its children.
This switch means that it is no longer necessary to pass changes up and down the tree when `ItemDetailsPage` is rendered more than once (aka inside the document properties sidebar).
An added benefit is that this change seems to have stopped whatever issue that was occasionally causing the application to hang after an item was saved.