feat: use special async select component for async data
This component provides the graphql api client to the data function. Config files should only use this client variable - DO NOT IMPORT THE CLIENT (the api client uses data from the config files, so it is not ininitalized yet).
This component also changes the way the async select works. It not only every requests one document at a time -- even when converting previous strings or ids into the value objects.