feat: add WIP sign-in page
This replaces the old sign-in page with the new sign-in page.
The new page does not yet handle authentication with username and password, but it can check usernames.
The new page still allows signing in with GitHub.
The legacy page is retained for the sign-up flow for GitHub. The legacy page has been modfied to reset authentication more often, which should fix issues with clients getting stuck on the registration page (for the GitHub team). A route has been added for the legacy sign-in page.
The splash screen redirects to the legacy route when the user profile is not part of the GitHub org, allowing the legacy sign-in page to handle the sign-up/registration flow.