feat(TeamPage): enable managing team name, members, and organizers
These changes allow setting the team name, adding new members (from the existing users), viewing the organizers, change user roles (if an organizer), and removing a user from the team (if an organizer). Users who are not a team's organizer will not be able to make changes, but they can see the other users in the team and which of those users are organizers.