feat(tiptap): debounce forced updates
This tiptap editor hook increments a counter on every editor transaction. This causes every component that relies on the editor to re-render. This is important for ensuring the editor is in sync with presence indicators and decorations.
However, re-rendering can be expensive. If the writer performs rapid changes such as holding the backspace key or typing extremely fast, the app can hang while it attempts to work its ways through all of the re-renders.
This commit debounces the counter increment by 30ms to ensure the editor does not hang.