feat!: use marks for comments instead of inline node
Because inline nodes are no longer used, comments must be rendered separately.
Comments are now rendered in a side panel.
This change will make it easier to add replies in the future.
This change fixes an issue related to lost comments after scrolling.
This change fixes an issue with not being able to comment at the bottom of the screen.
Comments are no longer automatically deleted when they are empty.