feat: provide current json to TipTap to pass to onDebouncedChange
This makes it possible to compare the current editor json with the current state json. Attempting to compare state within onDebouncedChange does not work because it uses a stale state from when the component was first created.
For currentJsonInState, pass null if the state has not yet been populated so that an undefined value in the state is not confused with the state not yet containing and values.