feat: support using the legacy item page
When the local config defines fields for a collection, prefer the local config. Custom components are only supported in the local config.
Note that this will eventually be removed. Custom components are primarily used for generating a print or web UI, and this functionality will be replaced with preview URLs.