feat: exclusively handle ydoc from server instead of creating it locally (#14)
* feat: depend exclusively on server-created ydoc

This removes all logic related to injecting data from the server into ydoc shared types. This now occurs on the server, which should prevent issues with multiple clients injecting data and overwriting each other (or duplicating values).

* feat: use key based on collection name and item id

This will help ensure fields are replaced when navigating directly between docs

* fix: wrap attempt to encode ydoc state as update in try catch

In case a client's ydoc is corrupted, it will fail to encode as an update. We let the client know that the doc was corrupted, and we mark the doc as not loading so the client may try again to save the doc.

* feat: use better language on unsaved changes prompt

* feat: populate doc array with `getYFields`

This ensures that the objects inside the array have the correct values, and that the values are represented the same way that they would be represented outside of the doc array.
This is also removes dependence on the deprecated `ydoc.getJSON()` method.

* feat: expose doc data in useFindDoc hook

* feat: only fetch yState and supplementary fields (yState already contains most data)

* feat: merge supplementary data with field data for previews