fix: access route-correct params during page slot transition
The $page store immediately reflects the new page params, which is problematic when the old page and the new page both still have content. In some cases, values may be undefined when they should not be undefined. To fix this, we must access the params by returning them from the page's load function.