feat: replace old user invite modal with new invite user dialog
This new dialog supports custom field descriptions, shows incrimented username and slug, and refetches the profiles list after a new user is invite. It prevents using an email address that is already used, and it refreshes the profiles list after invite. It also allows creating an account that is already deactivated.